Extract custom assertions of FeedViewControllerTests to a separate file
